+++ src/treecipe/src/ConfigurationService/ConfigurationService.ts
@@ -1,7 +1,7 @@
 
 import { VSCodeWorkspaceService } from '../VSCodeWorkspace/VSCodeWorkspaceService';
 import { IRecipeFakerService } from '../RecipeFakerService.ts/IRecipeFakerService';
-import { SnowfakeryFakerService } from '../RecipeFakerService.ts/SnowfakeryFakerService/SnowfakeryFakerService';
+import { SnowfakeryFakerService } from '../RecipeFakerService.ts/SnowfakeryRecipeFakerService/SnowfakeryRecipeFakerService';
 import { FakerJSRecipeFakerService } from '../RecipeFakerService.ts/FakerJSRecipeFakerService/FakerJSRecipeFakerService';
 import { SnowfakeryRecipeProcessor } from '../FakerRecipeProcessor/SnowfakeryRecipeProcessor/SnowfakeryRecipeProcessor';
 import { FakerJSRecipeProcessor } from '../FakerRecipeProcessor/FakerJSRecipeProcessor/FakerJSRecipeProcessor';
