+++ src/treecipe/src/RecipeFakerService.ts/FakerJSRecipeFakerService/tests/FakerJSRecipeFakerService.test.ts
@@ -3,12 +3,12 @@ import { FakerJSRecipeFakerService } from "../FakerJSRecipeFakerService";
 
 describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
 
-    let FakerJSRecipeFakerService = new FakerJSRecipeFakerService();
+    let fakerJSRecipeFakerService = new FakerJSRecipeFakerService();
 
     describe('getMapSalesforceFieldToFakerValue', () => {
 
-        let fakerService = new FakerJSRecipeFakerService();
-        const fieldTypeToNPMFakerMappings = fakerService.getMapSalesforceFieldToFakerValue();
+        // let fakerService = new FakerJSRecipeFakerService();
+        const fieldTypeToNPMFakerMappings = fakerJSRecipeFakerService.getMapSalesforceFieldToFakerValue();
         
         
         const placeholderForXMLMarkupDependentValue = 'GENERATED BY FIELD XML MARKUP';
@@ -121,7 +121,7 @@ describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
         test('given 1 tab argument, returns expected spaces', () => {
 
             const expectedTabs = 4; 
-            const actualTabResult = FakerJSRecipeFakerService.generateTabs(1);
+            const actualTabResult = fakerJSRecipeFakerService.generateTabs(1);
             const actualSpacesCount = actualTabResult.length;
             expect(actualSpacesCount).toEqual(expectedTabs);
 
@@ -137,7 +137,7 @@ describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
             const expectedRecipeValue = `faker.helpers.arrayElements(['apple','orange','banana'])`;
             const emptyRecordTypeApiToRecordTypeWrapperMap: Record<string, RecordTypeWrapper> = {};
             const fakeFieldApiName = "Fruit__c";
-            const actualFakerValue = FakerJSRecipeFakerService.buildMultiSelectPicklistRecipeValueByXMLFieldDetail(possibleChoices, emptyRecordTypeApiToRecordTypeWrapperMap, fakeFieldApiName);
+            const actualFakerValue = fakerJSRecipeFakerService.buildMultiSelectPicklistRecipeValueByXMLFieldDetail(possibleChoices, emptyRecordTypeApiToRecordTypeWrapperMap, fakeFieldApiName);
 
             expect(actualFakerValue).toBe(expectedRecipeValue);
         });
@@ -191,7 +191,7 @@ describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
 
             const emptyRecordTypeNameByRecordTypeNameToXRecordTypeWrapperMap: Record<string, RecordTypeWrapper> = {};
             const fakeFieldApiName = "CarModel__c";
-            const actualFakerValue = FakerJSRecipeFakerService.buildDependentPicklistRecipeFakerValue(expectedControllingValueToPicklistOptions, 
+            const actualFakerValue = fakerJSRecipeFakerService.buildDependentPicklistRecipeFakerValue(expectedControllingValueToPicklistOptions, 
                                                         emptyRecordTypeNameByRecordTypeNameToXRecordTypeWrapperMap, 
                                                         controllingField, 
                                                         fakeFieldApiName);
@@ -209,7 +209,7 @@ describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
             const expectedRecipeValue = `faker.helpers.arrayElement(['apple','orange','banana'])`;
             const emptyRecordTypeNameByRecordTypeNameToXRecordTypeWrapperMap: Record<string, RecordTypeWrapper> = {};
             const fakeFieldApiName = "Fruit__c";
-            const actualFakerValue = FakerJSRecipeFakerService.buildPicklistRecipeValueByXMLFieldDetail(possibleChoices, 
+            const actualFakerValue = fakerJSRecipeFakerService.buildPicklistRecipeValueByXMLFieldDetail(possibleChoices, 
                 emptyRecordTypeNameByRecordTypeNameToXRecordTypeWrapperMap, 
                                                                                                 fakeFieldApiName);
 
@@ -222,7 +222,7 @@ describe('FakerJSRecipeFakerService Shared Intstance Tests', () => {
 
         test('given expected list of OOTB object keys, mapping keys are found', () => {
             
-            const actualMappings = FakerJSRecipeFakerService.getOOTBObjectApiNameToFieldApiNameMap();
+            const actualMappings = fakerJSRecipeFakerService.getOOTBObjectApiNameToFieldApiNameMap();
 
             const expectedObjectsInMappings = [
                 'Account',
